fix: duplicate extra reducer
